+++ src/Controller/Admin/TaxonomyTermController.php
@@ -372,4 +372,58 @@ class TaxonomyTermController extends AbstractActionController
         $view->setVariable('count', $count);
         return $view;
     }
+
+    public function linkedValuesAction()
+    {
+        $taxonomyTerm = $this->api()->read('taxonomy_terms', $this->params('id'))->getContent();
+
+        $partialViewHelper = $this->viewHelpers()->get('partial');
+        $query = $this->params()->fromQuery();
+
+        $values = [];
+        foreach ($taxonomyTerm->linkedValues($query) as $linkedValue) {
+            $resource = $linkedValue->resource();
+            $property = $linkedValue->property();
+            $values[] = [
+                $partialViewHelper('taxonomy/admin/taxonomy-term/linked-values/resource', ['resource' => $resource]),
+                $partialViewHelper('taxonomy/admin/taxonomy-term/linked-values/property', ['property' => $property]),
+                $partialViewHelper('taxonomy/admin/taxonomy-term/linked-values/taxonomy-term', ['taxonomyTerm' => $linkedValue->valueResource()]),
+            ];
+        }
+
+        $response = [
+            'values' => $values,
+            'total' => $taxonomyTerm->linkedValuesCount(),
+        ];
+
+        return new JsonModel($response);
+    }
+
+    public function linkedValuesGridConfigAction()
+    {
+        $config = [
+            'columns' => [
+                $this->translate('Resource'),
+                $this->translate('Property'),
+                $this->translate('Taxonomy term'),
+            ],
+            'server' => [
+                'url' => $this->url()->fromRoute('admin/taxonomy-term-id', ['action' => 'linked-values'], [], true),
+            ],
+            'autoWidth' => false,
+            'pagination' => [
+                'limit' => $this->settings()->get('pagination_per_page', 25),
+                'summary' => false,
+                'buttonsCount' => 0,
+            ],
+            'language' => [
+                'pagination' => [
+                    'previous' => $this->translate('Previous'),
+                    'next' => $this->translate('Next'),
+                ],
+            ],
+        ];
+
+        return new JsonModel($config);
+    }
 }