+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show children count on term show page and in show-details sidebar
 - Modify admin and site advanced search forms to be able to search resources
   that links to a specific taxonomy term or its descendants.
+- On taxonomy term pages, show resources linked to term descendants too
 
 ## [0.5.1] - 2023-11-23